Add graphdriver.ErrDriverNotImplemented
If a driver returns this from any of its Differ methods we fall back
automatically to the default implementation rather than erroring
out. We will need this later in the aufs backend as we will only
sometimes be able to use a faster implementation (depending on e.g.
the layer depth).