beam: Fix double-close of fds in FdConn
We pass an *os.File instead of an fd int into FdConn, because
otherwise we have to call os.NewFile on the passed in Fd, which
is problematic as that sets a finalizer for that object which closes
the fd. This means the fd passed in will be closed during GC, which
is not expected.

Docker-DCO-1.1-Signed-off-by: Alexander Larsson <alexl@redhat.com> (github: alexlarsson)