beam: Make fd passing limit more robust
We pass a maxFds to receiveUnix, which lets us limit the oob data size.
This means the kernel will close any extra fds for us before creating
fds for the process. This protects us against filling up the process
file descriptor table.

We also pass extractFds the max and close any extra, because the size
limit is not perfect due to padding/alignment. For instance, on 64bit
arches the alignment of cmsgdata is 8, but that can fit two fds, so we
can't limit to 1 fd that way.

Docker-DCO-1.1-Signed-off-by: Alexander Larsson <alexl@redhat.com> (github: alexlarsson)