beam: Add simple framing system
This is needed for Send/Recieve to correctly handle borders between
the messages.

Also, in order for this to work i need to store recieved file
descriptors as they may be read before the correspondinf message.
So, we add a beam.BeamConn type to store this in and make the
operations methods on that instead.

Docker-DCO-1.1-Signed-off-by: Alexander Larsson <alexl@redhat.com> (github: alexlarsson)