Engine.Nuke destroys a docker engine and its filesystem state.
This is a convenience for test suites which create temporary engines and
need an easy way to remove them.

Note that Engine no longer needs to store filesystem state in the first
place, so Engine.Root should be deprecated. Once it is, Nuke should be
removed as well.

Docker-DCO-1.1-Signed-off-by: Solomon Hykes <solomon@docker.com> (github: shykes)